requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
alembic
psycopg2-binary
python-jose
passlib
bcrypt
pytest
pytest-cov
python-multipart
httpx